fix: Workspaces are build inconsistently
If a workspace was not rebuilt, it would not be written to the state file. This caused the project to be rebuilt on the next build, as it appeared to have never been built. The projects that then didn't have to be rebuilt on this run, would be missing from the state file and so on.