Adding `--ignore-dependencies`, `--fail-fast`, improve `--continue-on-error`, and improving CI report (#225)
Adding --ignore-dependencies and setting it as the default for running tests. Tests in one package cannot depend on the output of tests in another package, and there seems to be universal agreement on this. So it follows that tests can be run in full parallel.

Adding --fail-fast and fixing up --continue-on-error, so that jobs aren't killed part way through unless --fail-fast is set. When --continue-on-error is set, a failing job will not prevent subsequent jobs from running.

Improved the CI run report.

This adds depth information to the run order, as well as correcting as issue where the first item of a new level, was shown on the wrong level.

Using color to make the realtime output clearer for success and failures. As well as printing out the runtime per job, not the runtime of that job since the start of the whole run.

Also printing out the cumulative, saved and wall time in the final run report.