Make formatHeader handle deeply nested tree
When formatHeader function was called in a project with a deep structure
the calculation of how many minus-signs that should be printed in the
report header could become negative. This negative value was fed into
the string repetition method causing an exception to be thrown.

This change rewrites and hardens the formatHeader function.